som edit

+++ admin/modules/user/views/index.inc.php
@@ -7,14 +7,25 @@ $path = "modules/user/views/";
 $user_model = new UserModel;
 $user_type_model = new UserTypeModel;
 
-date_default_timezone_set("Asia/Bangkok");
-$d1=date("d");
-$d2=date("m");
-$d3=date("Y");
-$d4=date("H");
-$d5=date("i");
-$d6=date("s");
-$date="$d1$d2$d3$d4$d5$d6";
+
+
+
+    //---------------------ฟังก์ชั่นวันที่------------------------------------
+    date_default_timezone_set("Asia/Bangkok");
+    $d1=date("d");
+    $d2=date("m");
+    $d3=date("Y");
+    $d4=date("H");
+    $d5=date("i");
+    $d6=date("s");
+    $date="$d1$d2$d3$d4$d5$d6";
+    //---------------------------------------------------------------------
+
+
+    //-----------------ฟังก์ชั่นสุ่มตัวเลข----------------
+    $numrand = (mt_rand());
+    //-----------------------------------------------
+
 $target_dir = "../img_upload/user/";
 
 if(isset($_GET['id'])){
@@ -65,28 +76,70 @@ if(!isset($_GET['action'])){
         $data['user_username'] = trim($_POST['user_username']);
         $data['user_password'] = trim($_POST['user_password']);
 
+
+        
+        //------------------ฟังชั่นแก้ไขรูป--------------------
         if($_FILES['user_image']['name'] == ""){
-            $data['user_image'] = "";
+            $data['user_image'] = $_POST['user_image'];
         }else {
-            $target_file = $target_dir .$date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+            //---------เอาชื่อไฟล์เก่าออกให้เหลือแต่นามสกุล----------
+            $type = strrchr($_FILES['user_image']['name'],".");
+            //--------------------------------------------------
+            
+            //-----ตั้งชื่อไฟล์ใหม่โดยเอาเวลาไว้หน้าชื่อไฟล์เดิม---------
+            $newname = $date.$numrand.$type;
+            $path_copy=$path.$newname;
+            $path_link=$target_dir.$newname;
+            //-------------------------------------------------
+
+            $target_file = $target_dir .$date.$newname;
             $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
             // Check if file already exists
             if (file_exists($target_file)) {
-                $error_msg =  "ขอโทษด้วย. มีไฟล์นี้ในระบบแล้ว";
+                $error_msg =  "Sorry, file already exists.";
                 $check = false;
             }else if ($_FILES["user_image"]["size"] > 5000000) {
-                $error_msg = "ขอโทษด้วย. ไฟล์ของคุณต้องมีขนาดน้อยกว่า 5 MB.";
+                $error_msg = "Sorry, your file is too large.";
                 $check = false;
             }else if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" ) {
-                $error_msg = "ขอโทษด้วย. ระบบสามารถอัพโหลดไฟล์นามสกุล JPG, JPEG, PNG & GIF เท่านั้น.";
+                $error_msg = "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
                 $check = false;
             }else if (move_uploaded_file($_FILES["user_image"]["tmp_name"], $target_file)) {
-                $data['user_image'] = $date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+
+                //-----------------------------------
+                $data['user_image'] = $date.$newname;
+                //-----------------------------------
+
             } else {
-                $error_msg =  "ขอโทษด้วย. ระบบไม่สามารถอัพโหลดไฟล์ได้.";
+                $error_msg =  "Sorry, there was an error uploading your file.";
                 $check = false;
             } 
         }
+        //------------------------------------------------------------------------------
+
+
+        // if($_FILES['user_image']['name'] == ""){
+        //     $data['user_image'] = "";
+        // }else {
+        //     $target_file = $target_dir .$date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+        //     $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+        //     // Check if file already exists
+        //     if (file_exists($target_file)) {
+        //         $error_msg =  "ขอโทษด้วย. มีไฟล์นี้ในระบบแล้ว";
+        //         $check = false;
+        //     }else if ($_FILES["user_image"]["size"] > 5000000) {
+        //         $error_msg = "ขอโทษด้วย. ไฟล์ของคุณต้องมีขนาดน้อยกว่า 5 MB.";
+        //         $check = false;
+        //     }else if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" ) {
+        //         $error_msg = "ขอโทษด้วย. ระบบสามารถอัพโหลดไฟล์นามสกุล JPG, JPEG, PNG & GIF เท่านั้น.";
+        //         $check = false;
+        //     }else if (move_uploaded_file($_FILES["user_image"]["tmp_name"], $target_file)) {
+        //         $data['user_image'] = $date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+        //     } else {
+        //         $error_msg =  "ขอโทษด้วย. ระบบไม่สามารถอัพโหลดไฟล์ได้.";
+        //         $check = false;
+        //     } 
+        // }
 
         if($check == false){
             ?>  <script>  alert('<?php echo $error_msg; ?>'); window.history.back(); </script>  <?php
@@ -123,33 +176,82 @@ if(!isset($_GET['action'])){
         $data['user_password'] = trim($_POST['user_password']);
         $data['updateby'] = $login_user['user_id'];
 
+
+        
+        
+        //------------------ฟังชั่นแก้ไขรูป--------------------
         if($_FILES['user_image']['name'] == ""){
-            $data['user_image'] = $_POST['user_image_o'];
+            $data['user_image'] = $_POST['user_image'];
         }else {
-            $target_file = $target_dir .$date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+            //---------เอาชื่อไฟล์เก่าออกให้เหลือแต่นามสกุล----------
+            $type = strrchr($_FILES['user_image']['name'],".");
+            //--------------------------------------------------
+            
+            //-----ตั้งชื่อไฟล์ใหม่โดยเอาเวลาไว้หน้าชื่อไฟล์เดิม---------
+            $newname = $date.$numrand.$type;
+            $path_copy=$path.$newname;
+            $path_link=$target_dir.$newname;
+            //-------------------------------------------------
+
+            $target_file = $target_dir .$date.$newname;
             $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
             // Check if file already exists
             if (file_exists($target_file)) {
-                $error_msg =  "ขอโทษด้วย. มีไฟล์นี้ในระบบแล้ว";
+                $error_msg =  "Sorry, file already exists.";
                 $check = false;
             }else if ($_FILES["user_image"]["size"] > 5000000) {
-                $error_msg = "ขอโทษด้วย. ไฟล์ของคุณต้องมีขนาดน้อยกว่า 5 MB.";
+                $error_msg = "Sorry, your file is too large.";
                 $check = false;
             }else if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" ) {
-                $error_msg = "ขอโทษด้วย. ระบบสามารถอัพโหลดไฟล์นามสกุล JPG, JPEG, PNG & GIF เท่านั้น.";
+                $error_msg = "Sorry, only JPG, JPEG, PNG & GIF files are allowed.";
                 $check = false;
             }else if (move_uploaded_file($_FILES["user_image"]["tmp_name"], $target_file)) {
-                $data['user_image'] = $date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
 
-                $target_file = $target_dir . $_POST['user_image_o'];
-                if (file_exists($target_file)) {
-                    unlink($target_file);
+                //-----------------------------------
+                $data['user_image'] = $date.$newname;
+                //-----------------------------------
+
+                $target_file = $target_dir . $_POST["user_image"];
+                if($_POST["img_about_us_o"] != 'default.png'){
+                    if (file_exists($target_file)) {
+                        unlink($target_file);
+                    }
                 }
             } else {
-                $error_msg =  "ขอโทษด้วย. ระบบไม่สามารถอัพโหลดไฟล์ได้.";
+                $error_msg =  "Sorry, there was an error uploading your file.";
                 $check = false;
             } 
         }
+        //------------------------------------------------------------------------------
+
+
+        // if($_FILES['user_image']['name'] == ""){
+        //     $data['user_image'] = $_POST['user_image_o'];
+        // }else {
+        //     $target_file = $target_dir .$date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+        //     $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
+        //     // Check if file already exists
+        //     if (file_exists($target_file)) {
+        //         $error_msg =  "ขอโทษด้วย. มีไฟล์นี้ในระบบแล้ว";
+        //         $check = false;
+        //     }else if ($_FILES["user_image"]["size"] > 5000000) {
+        //         $error_msg = "ขอโทษด้วย. ไฟล์ของคุณต้องมีขนาดน้อยกว่า 5 MB.";
+        //         $check = false;
+        //     }else if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" ) {
+        //         $error_msg = "ขอโทษด้วย. ระบบสามารถอัพโหลดไฟล์นามสกุล JPG, JPEG, PNG & GIF เท่านั้น.";
+        //         $check = false;
+        //     }else if (move_uploaded_file($_FILES["user_image"]["tmp_name"], $target_file)) {
+        //         $data['user_image'] = $date.'-'.strtolower(basename($_FILES["user_image"]["name"]));
+
+        //         $target_file = $target_dir . $_POST['user_image_o'];
+        //         if (file_exists($target_file)) {
+        //             unlink($target_file);
+        //         }
+        //     } else {
+        //         $error_msg =  "ขอโทษด้วย. ระบบไม่สามารถอัพโหลดไฟล์ได้.";
+        //         $check = false;
+        //     } 
+        // }
         
         if($check == false){
             ?>  <script>  alert('<?php echo $error_msg; ?>'); window.history.back(); </script>  <?php
